Fix error propogation during switches
Before this fix, an error (such as `.STATICMODEVIOLATION`) in any of the branches would cause the final result to be the error (except for `.YulHalt`). The correct behaviour is for only that branch to be associated with the error, and not the whole result of the switch. This is now the behaviour.